2025-07-15-decaying-resolution-memory/pyproject.toml
[project]
name = "2025-07-15-decaying-resolution-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "baml-py>=0.201.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "tqdm>=4.67.1",
    "typing-extensions>=4.14.1",
]
